+++ filter_test.go
@@ -70,6 +70,7 @@ func TestPrinter_Filter(t *testing.T) {
 	t.Run("is passed the printer's config", func(t *testing.T) {
 		cfg := Config{
 			Indent:          []byte("--->"),
+			ZeroValueMarker: "*ZERO*",
 			RecursionMarker: "*LOOP*",
 		}