+++ CHANGELOG.md
@@ -9,6 +9,16 @@ The format is based on [Keep a Changelog], and this project adheres to
 [Keep a Changelog]: https://keepachangelog.com/en/1.0.0/
 [Semantic Versioning]: https://semver.org/spec/v2.0.0.html
 
+## [Unreleased]
+
+## Changed
+
+- `Print()` now accepts multiple values
+
+## Fixed
+
+- Fix panic when rendering unexported `time.Time` values
+
 ## [0.3.4] - 2019-11-05
 
 ## Changed