+++ CHANGELOG.md
@@ -9,7 +9,7 @@ The format is based on [Keep a Changelog], and this project adheres to
 [Keep a Changelog]: https://keepachangelog.com/en/1.0.0/
 [Semantic Versioning]: https://semver.org/spec/v2.0.0.html
 
-## [Unreleased]
+## [0.4.0] - 2020-05-05
 
 ## Added
 
@@ -110,6 +110,7 @@ The format is based on [Keep a Changelog], and this project adheres to
 [0.3.3]: https://github.com/dogmatiq/dapper/releases/tag/v0.3.3
 [0.3.4]: https://github.com/dogmatiq/dapper/releases/tag/v0.3.4
 [0.3.5]: https://github.com/dogmatiq/dapper/releases/tag/v0.3.5
+[0.4.0]: https://github.com/dogmatiq/dapper/releases/tag/v0.4.0
 
 [#6]: https://github.com/dogmatiq/dapper/issues/6
 [#7]: https://github.com/dogmatiq/dapper/issues/7